improvement(semantic/cfg): add the `ControlFlowBuilder`. It is a simple change, Before this, we had a lot of fields in our control flow graph that were only used during the build process. This PR aims to simplify the ControlFlowGraph.
