+++ crates/oxc_semantic/src/control_flow/builder.rs
@@ -1,123 +1,11 @@
-use oxc_span::CompactStr;
-use oxc_syntax::operator::{
-    AssignmentOperator, BinaryOperator, LogicalOperator, UnaryOperator, UpdateOperator,
+use super::{
+    AssignmentValue, AstNodeId, BasicBlockElement, CompactStr, ControlFlowGraph, EdgeType, Graph,
+    NodeIndex, PreservedExpressionState, PreservedStatementState, Register,
+    StatementControlFlowType,
 };
-use petgraph::{stable_graph::NodeIndex, Graph};
-
-use crate::AstNodeId;
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub enum Register {
-    Index(u32),
-    Return,
-}
-
-#[derive(Debug, Clone)]
-pub enum ObjectPropertyAccessBy {
-    PrivateProperty(CompactStr),
-    Property(CompactStr),
-    Expression(Register),
-}
-
-#[derive(Debug, Clone)]
-pub struct CollectionAssignmentValue {
-    pub id: AstNodeId,
-    pub elements: Vec<Register>,
-    pub spreads: Vec<usize>,
-    pub collection_type: CollectionType,
-}
-
-#[derive(Debug, Clone)]
-pub struct CalleeWithArgumentsAssignmentValue {
-    pub id: AstNodeId,
-    pub callee: Register,
-    pub arguments: Vec<Register>,
-    pub spreads: Vec<usize>,
-    pub call_type: CallType,
-}
-
-#[derive(Debug, Clone)]
-pub struct ObjectPropertyAccessAssignmentValue {
-    pub id: AstNodeId,
-    pub access_on: Register,
-    pub access_by: ObjectPropertyAccessBy,
-    pub optional: bool,
-}
-
-#[derive(Debug, Clone)]
-pub struct BinaryAssignmentValue {
-    pub id: AstNodeId,
-    pub a: Register,
-    pub b: Register,
-    pub operator: BinaryOp,
-}
-
-#[derive(Debug, Clone)]
-pub struct UpdateAssignmentValue {
-    pub id: AstNodeId,
-    pub expr: Register,
-    pub op: UpdateOperator,
-    pub prefix: bool,
-}
-
-#[derive(Debug, Clone)]
-pub struct UnaryExpressioneAssignmentValue(pub AstNodeId, pub UnaryOperator, pub Register);
-
-#[derive(Debug, Clone)]
-pub enum AssignmentValue {
-    ImplicitUndefined,
-    NotImplicitUndefined,
-}
-
-#[derive(Debug, Clone)]
-pub enum BinaryOp {
-    BinaryOperator(BinaryOperator),
-    LogicalOperator(LogicalOperator),
-    AssignmentOperator(AssignmentOperator),
-}
-
-#[derive(Debug, Clone)]
-pub enum CollectionType {
-    Array,
-    // Note: we do not currently track object names in objects.
-    Object,
-    JSXElement,
-    JSXFragment,
-    // doesn't use spreads
-    Class,
-    TemplateLiteral,
-}
-
-#[derive(Debug, Clone)]
-pub enum CallType {
-    New,
-    CallExpression,
-    // the callee is the yielded value, arguments are always empty
-    // spreads are always empty
-    Yield,
-    // spreads are always empty
-    TaggedTemplate,
-    // spreads are always empty
-    Import,
-}
-
-#[derive(Debug, Clone)]
-pub enum BasicBlockElement {
-    Unreachable,
-    Assignment(Register, AssignmentValue),
-    Throw(Register),
-    Break(Option<Register>),
-}
-
-#[derive(Debug, Clone)]
-pub enum EdgeType {
-    Normal,
-    Backedge,
-    NewFunction,
-}
 
 #[derive(Debug, Default)]
-pub struct ControlFlowGraph {
+pub struct ControlFlowGraphBuilder {
     pub basic_blocks: Vec<Vec<BasicBlockElement>>,
     // note: this should only land in the big box for all things that take arguments
     // ie: callexpression, arrayexpression, etc
@@ -146,28 +34,20 @@ pub struct ControlFlowGraph {
     pub after_throw_block: Option<NodeIndex>,
 }
 
-impl ControlFlowGraph {
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    /// # Panics
-    pub fn basic_block_by_index(&self, index: NodeIndex) -> &Vec<BasicBlockElement> {
-        let idx =
-            *self.graph.node_weight(index).expect("expected a valid node index in self.graph");
-        self.basic_blocks.get(idx).expect("expected a valid node index in self.basic_blocks")
-    }
-
-    /// # Panics
-    pub fn basic_block_by_index_mut(&mut self, index: NodeIndex) -> &mut Vec<BasicBlockElement> {
-        let idx =
-            *self.graph.node_weight(index).expect("expected a valid node index in self.graph");
-        self.basic_blocks.get_mut(idx).expect("expected a valid node index in self.basic_blocks")
+impl ControlFlowGraphBuilder {
+    pub fn build(self) -> ControlFlowGraph {
+        ControlFlowGraph { graph: self.graph, basic_blocks: self.basic_blocks }
     }
 
     /// # Panics
     pub fn current_basic_block(&mut self) -> &mut Vec<BasicBlockElement> {
-        self.basic_block_by_index_mut(self.current_node_ix)
+        let idx = *self
+            .graph
+            .node_weight(self.current_node_ix)
+            .expect("expected `self.current_node_ix` to be a valid node index in self.graph");
+        self.basic_blocks
+            .get_mut(idx)
+            .expect("expected `self.current_node_ix` to be a valid node index in self.graph")
     }
 
     #[must_use]
@@ -352,59 +232,4 @@ impl ControlFlowGraph {
             debug_assert_eq!(popped_2.unwrap().0, id);
         }
     }
-}
-
-pub enum StatementControlFlowType {
-    DoesNotUseContinue,
-    UsesContinue,
-}
-
-pub struct PreservedStatementState {
-    put_label: bool,
-}
-
-pub struct PreservedExpressionState {
-    pub use_this_register: Option<Register>,
-    pub store_final_assignments_into_this_array: Vec<Vec<Register>>,
-}
-
-#[must_use]
-fn print_register(register: Register) -> String {
-    match &register {
-        Register::Index(i) => format!("${i}"),
-        Register::Return => "$return".into(),
-    }
-}
-
-#[must_use]
-pub fn print_basic_block(basic_block_elements: &Vec<BasicBlockElement>) -> String {
-    let mut output = String::new();
-    for basic_block in basic_block_elements {
-        match basic_block {
-            BasicBlockElement::Unreachable => output.push_str("Unreachable()\n"),
-            BasicBlockElement::Throw(reg) => {
-                output.push_str(&format!("throw {}\n", print_register(*reg)));
-            }
-
-            BasicBlockElement::Break(Some(reg)) => {
-                output.push_str(&format!("break {}\n", print_register(*reg)));
-            }
-            BasicBlockElement::Break(None) => {
-                output.push_str("break");
-            }
-            BasicBlockElement::Assignment(to, with) => {
-                output.push_str(&format!("{} = ", print_register(*to)));
-
-                match with {
-                    AssignmentValue::ImplicitUndefined => {
-                        output.push_str("<implicit undefined>");
-                    }
-                    AssignmentValue::NotImplicitUndefined => output.push_str("<value>"),
-                }
-
-                output.push('\n');
-            }
-        }
-    }
-    output
 }
